chore(gateway): split aggregator tests under 2000-line limit and fix fmt
Split the 2140-line tests.rs into mod.rs + helpers.rs + tool_tests.rs +
gateway_tests.rs (each < 1000 lines). Run cargo fmt to fix formatting.
