+++ crates/dcc-mcp-gateway/src/gateway/aggregator/tests/mod.rs
@@ -1,0 +1,3 @@
+mod gateway_tests;
+pub(crate) mod helpers;
+mod tool_tests;
